knowledge-base-sediment/scripts/requirements.txt
pyyaml>=6.0
pypdf>=4.0
markdown-it-py>=3.0
pygments>=2.17
pytest>=8.0
